feat: move entity fetchers to db/users
