+++ app/src/app/api/tag-contacts/route.ts
@@ -2,6 +2,7 @@ import m from "@mailchimp/mailchimp_marketing"
 import { NextRequest } from "next/server"
 
 import { prisma } from "@/db/client"
+import { getAggregatedRecords } from "@/db/users"
 import { getAggregatedData } from "@/lib/api/eas/aggregated"
 import mailchimp from "@/lib/mailchimp"
 
@@ -79,168 +80,7 @@ const fetchRecords = async (): Promise<EntityRecords> => {
   }
 
   const records = await getAggregatedData()
-
-  const [citizen, badgeholder, gov_contribution, rf_voter] = await Promise.all([
-    prisma.userAddress.findMany({
-      where: {
-        AND: [
-          {
-            address: {
-              in: records.citizen,
-            },
-          },
-          {
-            OR: [
-              {
-                tag: {
-                  not: "citizen",
-                },
-              },
-              {
-                tag: null,
-              },
-            ],
-          },
-        ],
-      },
-      select: {
-        address: true,
-        user: {
-          select: {
-            emails: {
-              select: {
-                email: true,
-              },
-            },
-          },
-        },
-      },
-    }),
-    prisma.userAddress.findMany({
-      where: {
-        AND: [
-          {
-            address: {
-              in: records.badgeholder,
-            },
-          },
-          {
-            OR: [
-              {
-                tag: {
-                  not: "badgeholder",
-                },
-              },
-              {
-                tag: null,
-              },
-            ],
-          },
-        ],
-      },
-      select: {
-        address: true,
-        user: {
-          select: {
-            emails: {
-              select: {
-                email: true,
-              },
-            },
-          },
-        },
-      },
-    }),
-    prisma.userAddress.findMany({
-      where: {
-        AND: [
-          {
-            address: {
-              in: records.gov_contribution,
-            },
-          },
-          {
-            OR: [
-              {
-                tag: {
-                  not: "gov_contribution",
-                },
-              },
-              {
-                tag: null,
-              },
-            ],
-          },
-        ],
-      },
-      select: {
-        address: true,
-        user: {
-          select: {
-            emails: {
-              select: {
-                email: true,
-              },
-            },
-          },
-        },
-      },
-    }),
-    prisma.userAddress.findMany({
-      where: {
-        AND: [
-          {
-            address: {
-              in: records.rf_voter,
-            },
-          },
-          {
-            OR: [
-              {
-                tag: {
-                  not: "rf_voter",
-                },
-              },
-              {
-                tag: null,
-              },
-            ],
-          },
-        ],
-      },
-      select: {
-        address: true,
-        user: {
-          select: {
-            emails: {
-              select: {
-                email: true,
-              },
-            },
-          },
-        },
-      },
-    }),
-  ])
-
-  const result = {
-    citizen: citizen.map((c) => ({
-      address: c.address,
-      email: c.user.emails.at(-1)?.email ?? "",
-    })),
-    badgeholder: badgeholder.map((b) => ({
-      address: b.address,
-      email: b.user.emails.at(-1)?.email ?? "",
-    })),
-    gov_contribution: gov_contribution.map((g) => ({
-      address: g.address,
-      email: g.user.emails.at(-1)?.email ?? "",
-    })),
-    rf_voter: rf_voter.map((r) => ({
-      address: r.address,
-      email: r.user.emails.at(-1)?.email ?? "",
-    })),
-  }
+  const result = await getAggregatedRecords(records)
 
   return result
 }
