+++ app/src/lib/api/eas/aggregated.ts
@@ -1,12 +1,12 @@
 import { easClient } from "./client"
 
-type ResponseProps = {
+export type AggregatedDataResponse = {
   citizen: string[]
   badgeholder: string[]
   gov_contribution: string[]
   rf_voter: string[]
 }
-export const getAggregatedData = async (): Promise<ResponseProps> => {
+export const getAggregatedData = async (): Promise<AggregatedDataResponse> => {
   const records = await easClient("/entities/aggregated", {
     // NOTE: Do we need this? Not sure if our EAS Indexer is publiclly accessible as I see no authorisation anywhere
     // headers: {
